Global refactor for test functions
