+++ ydb/core/kqp/common/result_set_format/ut/kqp_formats_arrow_ut.cpp
@@ -2,14 +2,11 @@
 
 #include <util/string/hex.h>
 
+#include <ydb/core/kqp/common/result_set_format/ut/kqp_formats_ut_helpers.h>
 #include <ydb/core/kqp/common/result_set_format/kqp_formats_arrow.h>
-
-#include <ydb/library/formats/arrow/simple_builder/array.h>
-#include <ydb/library/formats/arrow/simple_builder/filler.h>
 #include <ydb/library/yverify_stream/yverify_stream.h>
-
 #include <ydb/public/lib/scheme_types/scheme_type_id.h>
-#include <yql/essentials/minikql/computation/mkql_block_reader.h>
+
 #include <yql/essentials/minikql/computation/mkql_value_builder.h>
 #include <yql/essentials/minikql/mkql_string_util.h>
 #include <yql/essentials/public/udf/arrow/defs.h>
@@ -20,7 +17,6 @@
 #include <library/cpp/type_info/tz/tz.h>
 
 using namespace NKikimr::NMiniKQL;
-using namespace NKikimr::NArrow;
 using namespace NYql;
 
 inline static constexpr size_t TEST_ARRAY_SIZE = 1 << 16;
@@ -782,7 +778,7 @@ void TestDataTypeConversion(arrow::Type::type arrowTypeId) {
         values.emplace_back(GetValueOfBasicType(type, i));
     }
 
-    auto array = NTestUtils::MakeArray(values, type);
+    auto array = MakeArray(values, type);
     UNIT_ASSERT_C(array->ValidateFull().ok(), array->ValidateFull().ToString());
     UNIT_ASSERT(array->length() == static_cast<i64>(values.size()));
 
@@ -809,7 +805,7 @@ void TestDataTypeConversion(arrow::Type::type arrowTypeId) {
                 auto val = NBinaryJson::SerializeToJson(values[i].AsStringRef());
                 UNIT_ASSERT(static_cast<TPhysicalType>(typedArray->Value(i)) == val);
             } else {
-                auto value = NTestUtils::ExtractUnboxedValue(array, i, type, context.HolderFactory);
+                auto value = ExtractUnboxedValue(array, i, type, context.HolderFactory);
                 AssertUnboxedValuesAreEqual(value, values[i], type);
             }
         } else {
@@ -843,7 +839,7 @@ void TestFixedSizeBinaryDataTypeConversion() {
         values.emplace_back(GetValueOfBasicType(type, i));
     }
 
-    auto array = NTestUtils::MakeArray(values, type);
+    auto array = MakeArray(values, type);
     UNIT_ASSERT_C(array->ValidateFull().ok(), array->ValidateFull().ToString());
     UNIT_ASSERT(array->length() == static_cast<i64>(values.size()));
 
@@ -882,7 +878,7 @@ void TestSingularTypeConversion() {
         values.emplace_back();
     }
 
-    auto array = NTestUtils::MakeArray(values, type);
+    auto array = MakeArray(values, type);
     UNIT_ASSERT_C(array->ValidateFull().ok(), array->ValidateFull().ToString());
     UNIT_ASSERT(array->length() == static_cast<i64>(TEST_ARRAY_SIZE));
 
@@ -1062,7 +1058,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(structType));
 
         auto values = context.CreateStructs(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, structType);
+        auto array = NFormats::MakeArray(values, structType);
 
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(array->length() == static_cast<i64>(values.size()));
@@ -1103,7 +1099,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(tupleType));
 
         auto values = context.CreateTuples(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, tupleType);
+        auto array = NFormats::MakeArray(values, tupleType);
         UNIT_ASSERT(array->ValidateFull().ok());
 
         UNIT_ASSERT(array->length() == static_cast<i64>(values.size()));
@@ -1143,7 +1139,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         Y_ABORT_UNLESS(NFormats::IsArrowCompatible(listType));
 
         auto values = context.CreateListOfJsons(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, listType);
+        auto array = NFormats::MakeArray(values, listType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
         UNIT_ASSERT(array->type_id() == arrow::Type::LIST);
@@ -1176,7 +1172,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         Y_ABORT_UNLESS(NFormats::IsArrowCompatible(listType));
 
         auto values = context.CreateOptionalListOfOptional(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, listType);
+        auto array = NFormats::MakeArray(values, listType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
         UNIT_ASSERT(array->type_id() == arrow::Type::LIST);
@@ -1218,7 +1214,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
     //     auto values = context.CreateVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+    //     auto array = NFormats::MakeArray(values, variantType);
     //     UNIT_ASSERT(array->ValidateFull().ok());
     //     UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
     //     UNIT_ASSERT(array->type_id() == arrow::Type::DENSE_UNION);
@@ -1268,7 +1264,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
     //     auto values = context.CreateOptionalVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+    //     auto array = NFormats::MakeArray(values, variantType);
     //     UNIT_ASSERT(array->ValidateFull().ok());
     //     UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
     //     UNIT_ASSERT(array->type_id() == arrow::Type::STRUCT);
@@ -1331,7 +1327,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
     //     auto values = context.CreateDoubleOptionalVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+    //     auto array = NFormats::MakeArray(values, variantType);
     //     UNIT_ASSERT(array->ValidateFull().ok());
     //     UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
     //     UNIT_ASSERT(array->type_id() == arrow::Type::STRUCT);
@@ -1403,7 +1399,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+        auto array = NFormats::MakeArray(values, variantType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
         UNIT_ASSERT(array->type_id() == arrow::Type::DENSE_UNION);
@@ -1461,7 +1457,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateOptionalVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+        auto array = NFormats::MakeArray(values, variantType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
         UNIT_ASSERT(array->type_id() == arrow::Type::STRUCT);
@@ -1531,7 +1527,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueToNativeArrowConversion) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateDoubleOptionalVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+        auto array = NFormats::MakeArray(values, variantType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT(static_cast<ui64>(array->length()) == values.size());
         UNIT_ASSERT(array->type_id() == arrow::Type::STRUCT);
@@ -1610,7 +1606,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueDoNotFitToArrow) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(dictType));
 
         auto values = context.CreateDictUtf8ToInterval(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, dictType);
+        auto array = NFormats::MakeArray(values, dictType);
         UNIT_ASSERT(array->ValidateFull().ok());
 
         UNIT_ASSERT(array->type_id() == arrow::Type::STRUCT);
@@ -1655,7 +1651,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueDoNotFitToArrow) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(dictType));
 
         auto values = context.CreateDictOptionalToTuple(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, dictType);
+        auto array = NFormats::MakeArray(values, dictType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT_EQUAL(static_cast<ui64>(array->length()), values.size());
         UNIT_ASSERT_EQUAL(array->type_id(), arrow::Type::STRUCT);
@@ -1710,7 +1706,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueDoNotFitToArrow) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(doubleOptionalType));
 
         auto values = context.CreateOptionalOfOptional(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, doubleOptionalType);
+        auto array = NFormats::MakeArray(values, doubleOptionalType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT_EQUAL(static_cast<ui64>(array->length()), values.size());
 
@@ -1764,7 +1760,7 @@ Y_UNIT_TEST_SUITE(DqUnboxedValueDoNotFitToArrow) {
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateLargeVariant(1000);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
+        auto array = NFormats::MakeArray(values, variantType);
         UNIT_ASSERT(array->ValidateFull().ok());
         UNIT_ASSERT_EQUAL(static_cast<ui64>(array->length()), values.size());
         UNIT_ASSERT_EQUAL(array->type_id(), arrow::Type::DENSE_UNION);
@@ -1786,8 +1782,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
 
         auto structType = context.GetStructType();
         auto values = context.CreateStructs(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, structType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, structType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, structType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, structType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], structType);
@@ -1801,8 +1797,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(tupleType));
 
         auto values = context.CreateTuples(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, tupleType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, tupleType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, tupleType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, tupleType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], tupleType);
@@ -1816,8 +1812,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(dictType));
 
         auto values = context.CreateDictUtf8ToInterval(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, dictType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, dictType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, dictType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, dictType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], dictType);
@@ -1831,8 +1827,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         Y_ABORT_UNLESS(NFormats::IsArrowCompatible(listType));
 
         auto values = context.CreateListOfJsons(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, listType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, listType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, listType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, listType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], listType);
@@ -1846,8 +1842,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         Y_ABORT_UNLESS(NFormats::IsArrowCompatible(listType));
 
         auto values = context.CreateOptionalListOfOptional(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, listType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, listType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, listType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, listType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], listType);
@@ -1861,8 +1857,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
     //     auto values = context.CreateVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, variantType);
-    //     auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, variantType, context.HolderFactory);
+    //     auto array = NFormats::MakeArray(values, variantType);
+    //     auto restoredValues = NFormats::ExtractUnboxedValues(array, variantType, context.HolderFactory);
     //     UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
     //     for (ui64 index = 0; index < values.size(); ++index) {
     //         AssertUnboxedValuesAreEqual(values[index], restoredValues[index], variantType);
@@ -1876,8 +1872,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(optionalVariantType));
 
     //     auto values = context.CreateOptionalVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, optionalVariantType);
-    //     auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, optionalVariantType, context.HolderFactory);
+    //     auto array = NFormats::MakeArray(values, optionalVariantType);
+    //     auto restoredValues = NFormats::ExtractUnboxedValues(array, optionalVariantType, context.HolderFactory);
     //     UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
     //     for (ui64 index = 0; index < values.size(); ++index) {
     //         AssertUnboxedValuesAreEqual(values[index], restoredValues[index], optionalVariantType);
@@ -1891,8 +1887,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
     //     UNIT_ASSERT(NFormats::IsArrowCompatible(doubleOptionalVariantType));
 
     //     auto values = context.CreateDoubleOptionalVariantOverStruct(100);
-    //     auto array = NFormats::NTestUtils::MakeArray(values, doubleOptionalVariantType);
-    //     auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, doubleOptionalVariantType, context.HolderFactory);
+    //     auto array = NFormats::MakeArray(values, doubleOptionalVariantType);
+    //     auto restoredValues = NFormats::ExtractUnboxedValues(array, doubleOptionalVariantType, context.HolderFactory);
     //     UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
     //     for (ui64 index = 0; index < values.size(); ++index) {
     //         AssertUnboxedValuesAreEqual(values[index], restoredValues[index], doubleOptionalVariantType);
@@ -1906,8 +1902,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, variantType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, variantType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, variantType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], variantType);
@@ -1921,8 +1917,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(optionalVariantType));
 
         auto values = context.CreateOptionalVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, optionalVariantType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, optionalVariantType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, optionalVariantType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, optionalVariantType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], optionalVariantType);
@@ -1936,8 +1932,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(doubleOptionalVariantType));
 
         auto values = context.CreateDoubleOptionalVariantOverTupleWithOptionals(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, doubleOptionalVariantType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, doubleOptionalVariantType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, doubleOptionalVariantType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, doubleOptionalVariantType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], doubleOptionalVariantType);
@@ -1951,8 +1947,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(dictType));
 
         auto values = context.CreateDictOptionalToTuple(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, dictType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, dictType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, dictType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, dictType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], dictType);
@@ -1966,8 +1962,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(doubleOptionalType));
 
         auto values = context.CreateOptionalOfOptional(100);
-        auto array = NFormats::NTestUtils::MakeArray(values, doubleOptionalType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, doubleOptionalType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, doubleOptionalType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, doubleOptionalType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], doubleOptionalType);
@@ -1981,8 +1977,8 @@ Y_UNIT_TEST_SUITE(ConvertUnboxedValueToArrowAndBack){
         UNIT_ASSERT(NFormats::IsArrowCompatible(variantType));
 
         auto values = context.CreateLargeVariant(1000);
-        auto array = NFormats::NTestUtils::MakeArray(values, variantType);
-        auto restoredValues = NFormats::NTestUtils::ExtractUnboxedValues(array, variantType, context.HolderFactory);
+        auto array = NFormats::MakeArray(values, variantType);
+        auto restoredValues = NFormats::ExtractUnboxedValues(array, variantType, context.HolderFactory);
         UNIT_ASSERT_EQUAL(values.size(), restoredValues.size());
         for (ui64 index = 0; index < values.size(); ++index) {
             AssertUnboxedValuesAreEqual(values[index], restoredValues[index], variantType);
